player2 formatting code

+++ player2/components/include/ring.h
@@ -13,21 +13,21 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 See the GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License along
-with player2; see the file COPYING.  If not, write to the Free Software
+with player2; see the file COPYING. If not, write to the Free Software
 Foundation, 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 
 The Player2 Library may alternatively be licensed under a proprietary
 license from ST.
 
 Source file name : ring.h
-Author :           Nick
+Author : Nick
 
 Definition of the pure virtual class defining the interface to a simple ring
 storage device.
 
-Date        Modification                                    Name
-----        ------------                                    --------
-11-Dec-03   Created                                         Nick
+Date Modification Name
+---- ------------ --------
+11-Dec-03 Created Nick
 
 ************************************************************************/
 
@@ -38,7 +38,7 @@ Date        Modification                                    Name
 
 typedef enum
 {
-	RingNoError     = 0,
+	RingNoError = 0,
 	RingNoMemory,
 	RingTooManyEntries,
 	RingNothingToGet
@@ -46,7 +46,7 @@ typedef enum
 
 //
 
-#define RING_NONE_BLOCKING  0
+#define RING_NONE_BLOCKING 0
 
 //
 
@@ -54,16 +54,16 @@ class Ring_c
 {
 	public:
 
-		RingStatus_t    InitializationStatus;
+		RingStatus_t InitializationStatus;
 
 		virtual ~Ring_c(void) {};
 
-		virtual RingStatus_t Insert(uintptr_t Value) = 0;
-		virtual RingStatus_t Extract(uintptr_t *Value,
-									 unsigned int   BlockingPeriod = RING_NONE_BLOCKING) = 0;
+		virtual RingStatus_t Insert(unsigned int Value) = 0;
+		virtual RingStatus_t Extract(unsigned int *Value,
+					     unsigned int BlockingPeriod = RING_NONE_BLOCKING) = 0;
 		virtual RingStatus_t Flush(void) = 0;
-		virtual bool         NonEmpty(void) = 0;
+		virtual bool NonEmpty(void) = 0;
 };
 
-typedef class Ring_c    *Ring_t;
+typedef class Ring_c *Ring_t;
 #endif
